+++ executor/executor_test.go
@@ -3,16 +3,15 @@ package executor
 import (
 	"testing"
 
-	"github.com/gobench-io/gobench/executor/option"
 	"github.com/gobench-io/gobench/logger"
 	"github.com/stretchr/testify/assert"
 )
 
 func TestStart(t *testing.T) {
-	opts := &option.Options{
+	opts := &Options{
 		AgentSock:    "/tmp/a1",
 		ExecutorSock: "/tmp/e1",
-		DriverPath:   "../driver/script/valid-dnt/valid-dnt.so",
+		DriverPath:   "./driver/script/valid-dnt/valid-dnt.so",
 		AppID:        1,
 	}
 	logger := logger.NewNopLogger()
